+++ mgraph.h
@@ -10,7 +10,7 @@ typedef struct m_graph M_GRAPH;
 M_GRAPH* m_graph_create(int, int);
 void m_graph_free(M_GRAPH*);
 void m_graph_print(M_GRAPH*);
-void m_graph_insert(M_GRAPH*, int, int, int);
+void m_graph_insert(M_GRAPH*, int, int, float);
 int m_graph_checkedge(M_GRAPH*, int, int);
 void m_graph_adj(M_GRAPH*, int);
 void m_graph_remove(M_GRAPH*, int, int);
